Add `Layout` component with sticky header (#88)
* imported Layout component

* removed container from index file

* updated headerSticky prop to true in docs

* Fix the build

* removed unnecessary div

* removed unnecessary div
