+++ layout/index.ts
@@ -1,4 +1,5 @@
-export * from './Layout'
+export * from './DocumentContext'
+export * from './NavContext'
 export * from './MDXLayout'
 export * from './MDXLayoutNav'
 export * from './MDXLayoutPagination'
